Added confirmation dialog when deleting collections
Master layout not supports a new section for additional scripts. If a page requires certain scripts that are not required site wide, they can import said scripts to be available only on that page.

For example, the scenes/collections/show view now brings in sweet alert plugin to implement a delete confirmation dialog. A custom javascript script is also brought in from public/js folder that implements sweet alert to the collection deletion button.