Added card components for items
Collection show / details view now shows cards for items it parents. In views/components there is a new folder for item-cards. One for cards with a image, one for cards without any images.

Keep in mind these cards are not completly function yet (link for "more" does not work) and we do not currently support image uploads. But the basic template is still functional.